wip refactoring packages

+++ src/main/java/br/com/cesarsicas/springstore/web/controller/merchant/MerchantProductsController.java
@@ -1,19 +1,17 @@
 package br.com.cesarsicas.springstore.web.controller.merchant;
 
-import br.com.cesarsicas.springstore.domain.product.ProductDto;
-import br.com.cesarsicas.springstore.domain.product.UpdateProductDto;
-import br.com.cesarsicas.springstore.domain.product_category.ProductCategoryDto;
-import br.com.cesarsicas.springstore.domain.product_category.ProductCategoryRepository;
-import br.com.cesarsicas.springstore.domain.product.ProductEntity;
-import br.com.cesarsicas.springstore.domain.product.ProductRepository;
-import br.com.cesarsicas.springstore.domain.user.UserEntity;
+import br.com.cesarsicas.springstore.web.model.ProductDto;
+import br.com.cesarsicas.springstore.data.product.ProductEntity;
+import br.com.cesarsicas.springstore.data.product.ProductRepository;
+import br.com.cesarsicas.springstore.web.model.UpdateProductDto;
+import br.com.cesarsicas.springstore.web.model.ProductCategoryDto;
+import br.com.cesarsicas.springstore.data.product_category.ProductCategoryRepository;
+import br.com.cesarsicas.springstore.data.user.UserEntity;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
@@ -45,15 +43,14 @@ public class MerchantProductsController {
 
         var category = productCategoryRepository.findById(productDto.category());
 
-        if (category.isPresent()){
+        if (category.isPresent()) {
             productRepository.save(new ProductEntity(productDto, category.get(), user));
             return ResponseEntity.ok().build();
-        }else {
+        } else {
             return ResponseEntity.badRequest().build(); //todo check error code
         }
     }
 
-    //todo only products of this merchant
     @Transactional
     @PutMapping
     public ResponseEntity update(@RequestBody @Valid UpdateProductDto updateProductDto,
@@ -61,7 +58,7 @@ public class MerchantProductsController {
 
         ProductEntity productEntity = productRepository.getReferenceById(updateProductDto.id());
 
-        if (!Objects.equals(productEntity.getUser().getId(), user.getId())){
+        if (!Objects.equals(productEntity.getUser().getId(), user.getId())) {
             return ResponseEntity.badRequest().build();
         }
 
@@ -75,14 +72,18 @@ public class MerchantProductsController {
     @GetMapping("/categories")
     public ResponseEntity list() {
         return ResponseEntity.ok(productCategoryRepository.findAll().stream().map(ProductCategoryDto::new));
-
     }
 
-    //todo only products of this merchant
     @DeleteMapping("/productId")
     @Transactional
-    public ResponseEntity delete(@PathVariable Long id) {
+    public ResponseEntity delete(@PathVariable Long id,
+                                 @AuthenticationPrincipal UserEntity user) {
         ProductEntity productEntity = productRepository.getReferenceById(id);
+
+        if (!Objects.equals(productEntity.getUser().getId(), user.getId())) {
+            return ResponseEntity.badRequest().build();
+        }
+
         productRepository.delete(productEntity);
         return ResponseEntity.ok().build();
     }
